requirements.txt
nltk==3.4.5
numpy==1.18.1
pandas==0.25.3
pkg-resources==0.0.0
praat-parselmouth==0.3.3
PyAudio==0.2.11
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
SpeechRecognition==3.8.1
